feat(a2a-tools): introduce agent tools module with validation and utility functions for agent configuration and API URL checks
