+++ sites/playground/server/src/a2a-tools/index.ts
@@ -1,0 +1,2 @@
+export { buildAgentTools, isAllowedAgentUrl } from './agent-tools.js';
+export type { PlaygroundAgentConfig } from './agent-tools.js';
